turn off signed git pushes?
I'm not sure why this would help, nor am I sure it did. I had weird
problems today with pushing to github, and it said that signed pushes
were not supported. I've had it configured this way to only sign pushes
when asked for ages and never had a problem - even a successful push
just 10 min. before!  After changing this, I got a different error that
it couldn't find the repo at all, then retried again and it worked.
Changing it for now just in case GH is doing something silly like asking
for signed pushes when it doesn't support it (the commits are still
signed though, which is fine).